UserProfile/requirements.txt
flask==2.0.1
flask-pymongo==2.3.0
werkzeug==2.0.1
pyjwt==2.1.0
python-dotenv==0.19.0